pyproject.toml
[tool.poetry]
name = "crm-system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django-dotenv = "^1.4.2"
django-filter = "^25.1"
mysqlclient = "^2.2.7"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.10"
celery = "^5.4.0"
redis = "^6.2.0"
django-celery-results = "^2.5.1"
drf-excel = "^2.5.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
coverage = "^7.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
